test-sms-email-data/requirements.txt
aiohttp
tqdm
python-dotenv
pymongo
